fitch the data in the quote component and build the react app
