+++ src/components/quote/quote.css
@@ -1,3 +1,4 @@
-h6 {
+h4 {
   font-family: 'Courier New', Courier, monospace;
+  margin-left: 2rem;
 }
